Fix: Use custom hook to search books
